Add portable mode
The portable mode means that all zed rather than using the global
config, cache, temp or support folders it stores everything as
subfolders of a single Zed data folder. The portable mode is activated
by setting the environment variable ZED_PORTABLE_DIR to what the user
wants the Zed data folder to be.

This is useful for a number of different cases:
* Quickly testing alternate configurations / extensions / setups without
  modifying the global default configuration.
* Keeping a zed executable including all configuration and supporting
  files on a removable media such as a USB stick or network drive, which
  can then be moved between different computers without impacting the
  running computer and bringing configuration along with it.
* Some corporate Windows environments prohibit installing extensions to
  AppData